:art: fixed linter and the corresponding errors
